BUGFIX: add --ci commandline parameter to tests to use `skipIf` on certain test classes.
- remove `threaded_test.py` - this was not being used.
- rename `test_coordinate` to `test_coordinate_creation`
- fix syntax errors in test_util

+++ podpac/core/coordinate/test/test_util.py
@@ -52,13 +52,13 @@ class TestCoordUtils(object):
         # numbers
         assert make_coord_value(10.5) == 10.5
         assert make_coord_value(10) == 10.0
-        assert make_coord_value(10L) == 10.0
+        assert make_coord_value(long(10)) == 10.0
         assert make_coord_value(np.array(10.5)) == 10.5
         assert make_coord_value(np.array([10.5])) == 10.5
 
         assert type(make_coord_value(10.5)) is float
         assert type(make_coord_value(10)) is float
-        assert type(make_coord_value(10L)) is float
+        assert type(make_coord_value(long(10))) is float
         assert type(make_coord_value(np.array(10.5))) is float
         assert type(make_coord_value(np.array([10.5]))) is float
 
@@ -88,13 +88,13 @@ class TestCoordUtils(object):
         # numbers
         assert make_coord_delta(10.5) == 10.5
         assert make_coord_delta(10) == 10.0
-        assert make_coord_delta(10L) == 10.0
+        assert make_coord_delta(long(10)) == 10.0
         assert make_coord_delta(np.array(10.5)) == 10.5
         assert make_coord_delta(np.array([10.5])) == 10.5
 
         assert type(make_coord_delta(10.5)) is float
         assert type(make_coord_delta(10)) is float
-        assert type(make_coord_delta(10L)) is float
+        assert type(make_coord_delta(long(10))) is float
         assert type(make_coord_delta(np.array(10.5))) is float
         assert type(make_coord_delta(np.array([10.5]))) is float
